backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
